requirements.txt
flask
Werkzeug
Pillow